decrease headers level
I used to put the title in a H1 header. Now there is the awesome feature that shows the file name at the top of the file. So i removed the H1 titles and now i decrease the level of all the headers so that They are not always shifted.